fix(cleanups): cleaning up home screen component since its almust at 1000 lines
