+++ apps/passport-client/new-components/screens/Home/hooks/useWindowWidth.ts
@@ -1,0 +1,18 @@
+import { useEffect, useState } from "react";
+
+export const useWindowWidth = (): number => {
+  const [windowWidth, setWindowWidth] = useState(window.innerWidth);
+  useEffect(() => {
+    const onResize = (): void => {
+      setWindowWidth(window.innerWidth);
+    };
+
+    window.addEventListener("resize", onResize);
+
+    return () => {
+      window.removeEventListener("resize", onResize);
+    };
+  }, []);
+
+  return windowWidth;
+};
